redfish-emulator/requirements.txt
requests
gevent==20.5.0
grequests
randmac
pyyaml>=5.3.1
